Added swagger annotation in AdminRestController, CabinetRController, FeedbackRequestController
